AIML Projects/Img Anz/requirements.txt
Pillow
google-generativeai
streamlit
